+++ app/controllers/results_controller.rb
@@ -7,17 +7,11 @@ class ResultsController < ApplicationController
   end
 
   def students
-    @students = User.active
-                    .includes(room: :suite)
-                    .where(role: %w(student rep), college: College.current)
-                    .where.not(room_assignments: { room_id: nil })
-                    .order(:last_name)
+    @students = StudentsWithRoomsAssignedQuery.call
   end
 
   def export
-    s = User.active.where(role: %w(student rep), college: College.current)
-            .includes(:draw, :room, group: %i(lottery_assignment suite leader))
-            .order(:last_name)
+    s = ResultsQuery.call
     a = %I[username email student_id last_name first_name draw_name intent
            group_name lottery_number building_name suite_number room_number]