+++ app/models/group.rb
@@ -41,6 +41,7 @@ class Group < ApplicationRecord # rubocop:disable ClassLength
 
   delegate :number, to: :lottery_assignment, prefix: :lottery, allow_nil: true
   delegate :number, to: :suite, prefix: :suite, allow_nil: true
+  delegate :building_name, to: :suite, allow_nil: true
 
   enum status: %w(open closed finalizing locked)