+++ spec/models/group_spec.rb
@@ -31,6 +31,7 @@ RSpec.describe Group, type: :model do
     it { is_expected.to validate_presence_of(:transfers) }
     it { is_expected.not_to allow_value(-1).for(:transfers) }
     it { is_expected.to delegate_method(:suite_number).to(:suite).as(:number) }
+    it { is_expected.to delegate_method(:building_name).to(:suite) }
     it do
       is_expected.to delegate_method(:lottery_number).to(:lottery_assignment)
                                                      .as(:number)